:recycle: refactor : put comment component into a folder
